Change Buttons in the Designer Toolbar.
Change buttons in DesignerToolBar so that the Blocks button is grayed
out when in the blocks editor and the Designer button (formally the
"UI" button) is grayed out when in the designer. The Wireless Button
is now called "Connect to Companion" and is only shown when in the
blocks editor.

Change-Id: Id724eda55b23e52212b020cbcbe3297f13a3b1e6