Incremental Yail Update to the Companion
Once connected to the Companion, blocks are updated as they are
modified instead of sending the whole app to the phone on each
change. Errors evaluating the Yail in the Companion to flagged back to
the block that caused the problem and a red error triangle is added to
the top level block.

Change-Id: If79e04aa85b904a4da40b3e4e91fa11204b7167c