Deliver blockly-all.js in UTF-8 instead of UTF-16.
Arrange for App Engine server to provide the character set of UTF-8
for blockly-all.js in the Content-Type header. By default the browser
guesses the character set based on looking at the first "N"
characters.  Because Chinese was not used until far into the file, the
browser would incorrectly guess ISO-8859-1. UTF-16 worked because it
is a two byte encoding for *all* characters into the file, so the
browser was able to guess UTF-16 from the first N.

Also add the path of the closure library explicitly to the Plover
config file for when we actually use the compiler.

Change-Id: Ie410a6c42073188f07cc80f6dfca996f6b979c8c