Let the Companion Display Assets.
Arrange to fetch assets from the APK file in the Companion so it can
have a UI with images etc. Add a function to the PhoneStatus block
used to declare that the UI is loaded and assets should now be loaded
from the SD Card (because they would be loaded by the App being
developed!).

Change-Id: Ib2f407d97db9ac2fba7c7e242e3a974f2590b5f2