Added support for Tweets with Images, uploaded to TwitPic.
This also requires the user to input an API key given by
registering their app with TwitPic.

SetStatus and SetStatusWithImage have been renamed to
Tweet and TweetWithImage, as 'Tweet' is effectively a
verb and more intuitive than 'SetStatus'

TweetWithImage also tweets only the text if the image
is not found or is invalid.

Updated version numbers of the twitter component.

Change-Id: I6f2eace3468e6af229d326a8d739acef0f924419