Updates To Companion Connectivity and Project Switching.
Various fixes to permit Yail updates to the Companion while in the
designer as well as the blocks editor. Cleanup state when "disconnect
companion" is selected, so a fresh companion can be started and
connected to without reloading the project.

Added some try/catch blocks to deal with errors caused by race
conditions between blockly creation/destruction and the 2 second
timer. These changes will be removed when we re-factor the timer code
to run completely in blockly.

Change-Id: I7a0431d137097c38dde03104c94910b8fc3329a2