Shorten BlocksEditor.jar codebase.
Shorten the codebase in the jnlp file used to start the Blocks Editor.
We have removed the encrypted project information normally used for
authentication because we do not require authentication to download the
Blocks Editor (it is a public file) and having the encrypted project
information made the codebase attribute of the jnlp file sufficiently
long that it broke some implementations of Java.

Change-Id: I707f720a9f05fb0ea7097712643224eb75e27eae