Add Support for connecting to a device via USB Cable.
There is a new “Connect To” menu choice of “Connect via USB.” As its
name implies, you can use this to connect via a USB Cable. You still
have to install the Companion on the device, but you don’t have to
start it manually (it doesn’t hurt if you do). The connect menu choice
will start the Companion for you.

Updated aiDaemon (you will need a new setup package) to properly close
file descriptors prior to invoking “adb.” This means that you can kill
and restart aiDaemon without difficulty.

You can now use the Emulator even if you have a phone plugged
in. Previously this led to the blocks not being properly loaded to the
Companion in the emulator.

Change-Id: I9d21e01ef1df8b955d60a8472ce4be97a6e60515